Issue #702 - Force PayPalButtons reload
The closures created for PayPal callbacks need to
get updated in the PayPal React Context. This change
forces those updates using:

 - forceReRender parameter of PayPalButtons
 - sharing PayPalScriptProvider between frequency buttons
 and PayPal buttons
